+++ src/structures/GoalPost.java
@@ -27,7 +27,8 @@ public class GoalPost extends CircularShape {
 	 * Returns the description of the GoalPost.
 	 */
 	public String toString() {
-		return "goal(" + center.x + ", " + center.y + ", " + radius + ")";
+		return "goal(" + (int) center.x + ", " + (int) center.y + ", " + radius
+				+ ")";
 	}
 	
 }
